# openapi/LupiraMtgApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Lupira MTG API",
    "description": "Backend for the Lupira MTG mobile app — Magic: The Gathering card metadata, scan-based recognition, and per-user collection management. Authenticate with a Bearer token issued by Authentik (OIDC).",
    "version": "v1"
  },
  "paths": {
...
    "/cards": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "List or search functionally distinct cards (one entry per Oracle ID).",
        "description": "Returns one row per Oracle ID — i.e. functionally distinct cards. Each row carries\noracle-level fields (name, type line, oracle text, color identity, mana cost, CMC)\nplus a representative thumbnail picked deterministically (English, latest non-foil\nprinting with an image).\n\n**Filters** (all optional, AND'd together):\n- `q` — fuzzy match on card name (Postgres `pg_trgm`).\n- `set` — keep only cards with a printing in this set code (e.g. `m21`).\n- `rarity` — `common|uncommon|rare|mythic|special`.\n- `color` — single color identity letter (`W|U|B|R|G`).\n- `colors` — comma-separated multi-color identity (AND), e.g. `colors=W,U` keeps\n  only cards whose color identity contains every listed letter.\n- `type` — fuzzy match on the recomposed type line (e.g. `type=goblin warrior`).\n- `cmc` — exact converted mana cost.\n- `cmcMin` / `cmcMax` — CMC range bounds (inclusive).\n- `power` / `toughness` — exact match (strings, since `*` and `1+1` are valid).\n\n**Sort**: `sort=name|cmc|releasedAt|rarity|relevance`. Defaults to `relevance` when\n`q` is provided, otherwise `name`. `order=asc|desc` (relevance is always best-first).\n\n**Pagination**: `take` 1–100 (default 25), `skip` for paging.\n\nUse `GET /cards/{oracleId}/printings` to drill into printings for a specific card.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "set",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "color",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "colors",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rarity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cmc",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "cmcMin",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "cmcMax",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "float"
            }
          },
          {
            "name": "power",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toughness",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/cards/{oracleId}": {
      "get": {
        "tags": [
          "Cards"
        ],
        "summary": "Get a single card (oracle-level) by Oracle ID.",
        "description": "Returns the oracle-level view of a card: name, type line, oracle text, color identity,\nand a representative thumbnail. Does **not** return printing-specific data such as set,\ncollector number, or prices — see `GET /cards/{oracleId}/printings`.",
        "parameters": [
          {
            "name": "oracleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/sets": {
      "get": {
        "tags": [
          "Sets"
        ],
        "summary": "Browse the catalog of Magic sets.",
        "description": "Returns paginated set metadata. Use this to power a set picker (the `?set=`\nfilter on `GET /cards` takes the same `code` returned here).\n\n`setType` filters by Scryfall set type (`core`, `expansion`, `masters`,\n`commander`, `funny`, …).\n`sort` ∈ `releasedAt|code|name`, default `releasedAt`.\n`order` ∈ `asc|desc`, default `desc` (newest first).\n`take` 1–200 (default 50). `skip` for paging.",
        "parameters": [
          {
            "name": "setType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/sets/{code}": {
      "get": {
        "tags": [
          "Sets"
        ],
        "summary": "Get one set by its lower-case code.",
        "description": "Returns set metadata plus a presigned URL for the set icon (if cached locally).",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardFaceResponse": {
        "required": [
          "faceIndex",
          "name"
        ],
        "type": "object",
        "properties": {
          "faceIndex": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "manaCost": {
            "type": [
              "null",
              "string"
            ]
          },
          "typeLine": {
            "type": [
              "null",
              "string"
            ]
          },
          "oracleText": {
            "type": [
              "null",
              "string"
            ]
          },
          "power": {
            "type": [
              "null",
              "string"
            ]
          },
          "toughness": {
            "type": [
              "null",
              "string"
            ]
          },
          "images": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CardImageUrls"
              }
            ]
          }
        }
      },
      "CardImageUrls": {
        "type": "object",
        "properties": {
          "normal": {
            "type": [
              "null",
              "string"
            ]
          },
          "artCrop": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CardListResponse": {
        "required": [
          "results",
          "total"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "CardResponse": {
        "required": [
          "oracleId",
          "name",
          "typeLine",
          "colorIdentity",
          "layout"
        ],
        "type": "object",
        "properties": {
          "oracleId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "typeLine": {
            "type": "string"
          },
          "oracleText": {
            "type": [
              "null",
              "string"
            ]
          },
          "colorIdentity": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "manaCost": {
            "type": [
              "null",
              "string"
            ]
          },
          "cmc": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "float"
          },
          "power": {
            "type": [
              "null",
              "string"
            ]
          },
          "toughness": {
            "type": [
              "null",
              "string"
            ]
          },
          "layout": {
            "type": "string"
          },
          "thumbnail": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CardImageUrls"
              }
            ]
          },
          "printingCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "faces": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CardFaceResponse"
            }
          }
        }
      },
...
      "SetListResponse": {
        "required": [
          "results",
          "total"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "SetResponse": {
        "required": [
          "code",
          "name",
          "setType"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "setType": {
            "type": "string"
          },
          "releasedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "cardCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "iconUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "PoC device token from `POST /me/register`. Send as `Authorization: Bearer lmtg_<token>`. Replaced by Authentik OIDC tokens in the future Path C migration (see plan).",
        "scheme": "bearer"
      }
    }
  },
...
}